Create README.md
Except for the image gallery*, our website, madisonwest71.com is a static web page. It is built using common, free development tools including the Bootstrap 5 framework with the help of https://bootstrapstudio.io/. It is to be hosted on Netlify with a Namecheap domain name. To keep the madisonwest71.com domain name, payments of 13.48/year will be continue through namecheap.com. 

*The image gallery is built using jAlbum. I own my own copy of jAlbum. Any future website development could also use this but it is not free; there are many other image gallery tools that can be used.